Some refactoring, moving utility classes to common and renderers to renderer packages
