+++ api/src/main/java/org/openmrs/module/mksreports/patientsummary/GenericPatientSummary.java
@@ -19,6 +19,7 @@ import java.util.HashMap;
 import java.util.Properties;
 
 import org.openmrs.api.context.Context;
+import org.openmrs.module.mksreports.common.Helper;
 import org.openmrs.module.reporting.dataset.definition.EncounterAndObsDataSetDefinition;
 import org.openmrs.module.reporting.report.ReportDesign;
 import org.openmrs.module.reporting.report.definition.ReportDefinition;
