+++ api/src/main/java/org/openmrs/module/mksreports/renderer/XlsReportRenderer2.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.mksreports.patientsummary;
+package org.openmrs.module.mksreports.renderer;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.logging.Log;
@@ -7,6 +7,7 @@ import org.apache.poi.poifs.filesystem.POIFSFileSystem;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
 import org.openmrs.annotation.Handler;
+import org.openmrs.module.mksreports.common.ExcelBuilder2;
 import org.openmrs.module.reporting.common.Localized;
 import org.openmrs.module.reporting.common.ObjectUtil;
 import org.openmrs.module.reporting.dataset.DataSet;
